Update 02-SDI.py
In the SDI script I include two work-arounds in the Penetrability sub-index to include USFS trail data and dirt roads.  In the first iteration I manually merged the trail shapefile with the speed cat 7 & 8 subset from the streets shapefile.  In the second iteration I started coding the merge into python but am running into issues.  The changes are in lines 316-329.